pyproject.toml
[project]
name = "pictoria"
version = "0.1.0"
description = "Your Personal Image Gallery"
authors = [{ name = "Jannchie", email = "jannchie@gmail.com" }]
dependencies = [
    "fastapi>=0.115.5",
    "rich>=13.9.4",
    "wdtagger[gpu]>=0.11.1",
    "uvicorn>=0.32.0",
    "httpx>=0.27.2",
    "diffusers>=0.31.0",
    "openai>=1.55.1",
    "sqlmodel>=0.0.22",
    "watchdog>=6.0.0",
    "alembic>=1.14.0",
    "python-multipart>=0.0.17",
    "torch==2.5.1+cu124",
    "torchvision==0.20.1+cu124",
    "sqlalchemy>=2.0.36",
    "orjson>=3.10.12",
    "transformers>=4.46.3",
    "sentencepiece>=0.2.0",
    "protobuf>=5.28.3",
    "image-gen-aux",
    "pillow-avif-plugin>=1.4.6",
    "colorthief>=0.2.1",
    "datasets>=3.1.0",
    "sqlite-vec>=0.1.6",
    "accelerate>=1.2.0",
]

requires-python = ">=3.11"

[tool.black]
line-length = 120

[[tool.uv.index]]
name = "default"
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
image-gen-aux = { git = "https://github.com/asomoza/image_gen_aux.git" }

[dependency-groups]
dev = ["matplotlib>=3.9.3"]

[tool.pyright]
reportAssignmentType = "none"
reportArgumentType = "none"

[tool.ruff]
select = ["ALL"]
exclude = ["alembic"]
ignore = [
    "PGH",
    "RUF003",
    "BLE001",
    "ERA001",
    "FIX002",
    "TD002",
    "TD003",
    "D",
    "A004",
    "ANN201",
    "B008",
    "FAST002",
]
line-length = 120

[tool.uv]
package = true
